% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AF_vectors.R
\name{calc_AF_vectors}
\alias{calc_AF_vectors}
\title{Calculate Allele Frequency Change Vectors from a VCF}
\usage{
calc_AF_vectors(
  vcf = NULL,
  window_size = 200,
  popmap = NULL,
  vectors = NULL,
  n_cores = 1,
  normalise = TRUE,
  end_cutoff = window_size,
  null_perms = NULL
)
}
\arguments{
\item{vcf}{A vcf stored as a vcfR object}

\item{window_size}{Integer value describing how many SNPs to include per window}

\item{popmap}{A two column data.frame, column 1 lists individuals and column 2 population assignment}

\item{vectors}{A list object that descibes each vector. Each list element should be a character vector with two strings, each corresponding to a population in the popmap}

\item{n_cores}{number of cores to run with mclapply. If n_cores=1, runs with lapply instead.}

\item{normalise}{Boolean to determine whether allele frequencies should be normalised. Normalisation is essential for vector analysis, but setting normalise=FALSE permits comparison with the raw allele frequencies}

\item{end_cutoff}{Numeric value describing how many SNPs the last window can have. By default, this equals the window_size, such that the last window must have the same number of SNPs as all windows. This is highly recommended.}

\item{null_perms}{Integer value describing how many null permutations to run. When set to NULL, function returns observed allele frequency vectors. When set to any integer value, the function returns randomised permutated vectors that are used for estimation of the null distribution.}
}
\value{
A list object where each element corresponds to an m (number of SNPs per window) x n (number of populations) matrix of allele frequency change vectors
}
\description{
Takes as input a population map (popmap) and a VCF (vcfR object) and returns an m x n matrix, where m = the number of SNPs and n = the number of populations
}
