% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_analyse_vectors.R
\name{eigen_analyse_vectors}
\alias{eigen_analyse_vectors}
\title{Eigen Analysis of Allele Frequency Change Vector Matrix}
\usage{
eigen_analyse_vectors(vector_input)
}
\arguments{
\item{vector_input}{A matrix of allele frequency change vectors}
}
\value{
A list object with three elements detailing the results: "eigenvals" contains the eigenvalue distribution, "eigenvecs" contains the eigenvectors, "A_matrix" contains the projected SNP scores for each eigenvector
}
\description{
Performs eigen decomposition of a single matrix of normalised allele frequency change vectors. Can be run across the output of calc_AF_vectors() by combining with lapply()
}
