% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_pvals.R
\name{eigen_pvals}
\alias{eigen_pvals}
\title{Calculate associated p-values for AF-vapeR eigenvalues}
\usage{
eigen_pvals(eigen_res, null_vectors)
}
\arguments{
\item{eigen_res}{A list of results from eigen_analyse_vectors(). This can be easily created by combining eigen_analyse_vectors with lapply. Each element of the input list should correspond to the eigen decomposition results for a given genome window.}

\item{null_vectors}{A list of NULL allele frequency change vector matrices as output by calc_AF_vectors() with null_perms set to a numeric value. This is only necessary if wanting to calculate and plot p.values.}
}
\value{
Matrix of p-values for each window (rows) for each eigenvector (columns)
}
\description{
Takes a list containing results from eigen_analyse_vectors() and calculates p-values based on null vectors.
}
