% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenval_plot.R
\name{eigenval_plot}
\alias{eigenval_plot}
\title{Plot Windowed Eigenvalues Along a Chromosome}
\usage{
eigenval_plot(
  eigen_res = NULL,
  cutoffs = NULL,
  null_vectors = NULL,
  plot.pvalues = FALSE
)
}
\arguments{
\item{eigen_res}{A list of results from eigen_analyse_vectors(). This can be easily created by combining eigen_analyse_vectors with lapply. Each element of the input list should correspond to the eigen decomposition results for a given genome window.}

\item{cutoffs}{Numeric vector of alpha thresholds for null distribution cutoffs. These must be equal in length to the number of eigenvalues, such that the first element is the threshold for eigenvalue 1, the second for eigevalue 2 e.t.c.}

\item{null_vectors}{A list of NULL allele frequency change vector matrices as output by calc_AF_vectors() with null_perms set to a numeric value. This is only necessary if wanting to calculate and plot p.values.}

\item{plot.pvalues}{Boolean flag as to whether or not to plot p.values instead of summed eigenvalues.}
}
\value{
A list of ggplot figures, where each element corresponds to figure for each eigenvector.
}
\description{
Takes as input a list of results from eigen_analyse_vectors() and plots all results along a chromosome. The chromosome positions are derived from the names() variable of the eigen_res input, which should be in the standard genome interval format of chrX:X-Y
}
