% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenval_plot_genome.R
\name{eigenval_plot_genome}
\alias{eigenval_plot_genome}
\title{Plot Windowed Eigenvalues For A List of Chromosomes}
\usage{
eigenval_plot_genome(
  eigen_res_list = NULL,
  cutoffs = NULL,
  null_vectors = NULL,
  plot.pvalues = F
)
}
\arguments{
\item{eigen_res_list}{A list of lists containing results from eigen_analyse_vectors(). The length() of the eigen_res_list input should be equal to the number of chromosomes, such that each element of the input is itself a list derived from eigen_analyse_vectors()}

\item{cutoffs}{Numeric vector of alpha thresholds for null distribution cutoffs. These must be equal in length to the number of eigenvalues, such that the first element is the threshold for eigenvalue 1, the second for eigevalue 2 e.t.c.}

\item{null_vectors}{A list of NULL allele frequency change vector matrices as output by calc_AF_vectors() with null_perms set to a numeric value. This is only necessary if wanting to calculate and plot p.values.}

\item{plot.pvalues}{Boolean flag as to whether or not to plot p.values instead of summed eigenvalues.}
}
\value{
A list of ggplot figures, where each element corresponds to figure for each eigenvector.
}
\description{
Takes as input a list of lists containing results from eigen_analyse_vectors() and plots all results. The chromosome positions are derived from the names() variable of the eigen_res input, which should be in the standard genome interval format of chrX:X-Y.
}
