% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_eigen_res.R
\name{merge_eigen_res}
\alias{merge_eigen_res}
\title{Merge AF-vapeR Eigen Decomposition Results From Multiple Chromosomes}
\usage{
merge_eigen_res(eigen_res_list)
}
\arguments{
\item{eigen_res_list}{A list of results from eigen_analyse_vectors(). This can be easily created by combining eigen_analyse_vectors with lapply. Each element of the input list should correspond to the eigen decomposition results for a given genome window.}
}
\value{
Single list containing eigen analysis results for all chromosomes/scaffolds
}
\description{
Takes a list of multiple
}
