% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif_eigen_windows.R
\name{signif_eigen_windows}
\alias{signif_eigen_windows}
\title{Identify Windows with Parallel Evolution Evidence Based On Significant Eigenvalues}
\usage{
signif_eigen_windows(eigen_res = NULL, cutoffs = NULL)
}
\arguments{
\item{eigen_res}{A list of results from eigen_analyse_vectors(). This can be easily created by combining eigen_analyse_vectors with lapply. Each element of the input list should correspond to the eigen decomposition results for a given genome window.}

\item{cutoffs}{Numeric vector of alpha thresholds for null distribution cutoffs. These must be equal in length to the number of eigenvalues, such that the first element is the threshold for eigenvalue 1, the second for eigevalue 2 e.t.c.}
}
\value{
A list of character vectors, where each element of the list represents the significant windows associated with each eigenvalue.
}
\description{
Uses the results from eigen_analyse_vectors() and cutoffs derived from find_null_cutoffs() to return a list of significant windows associated with each eigenvector. Windows are associated with the highest order eigenvector for which their summed eigenvalues exceed the significance threshold.
}
