% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_eigenvals.R
\name{sum_eigenvals}
\alias{sum_eigenvals}
\title{Sum Eigenvalues Across eigen_analyse_vector Results}
\usage{
sum_eigenvals(eigen_res)
}
\arguments{
\item{eigen_res}{A single matrix eigen analysis results as output by eigen_analyse_vectors()}
}
\value{
A matrix of summed eigenvalues
}
\description{
This function takes the output the eigen analysis and returns a matrix of summed values where each eigenvalue is the sum of itself and the eigenvalues that preceed it. This summed value is what is plotted and tested for significance. This function can be run over a list of eigen_analyse_vectors with lapply().
}
