% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfR2AF.R
\name{vcfR2AF}
\alias{vcfR2AF}
\title{Calculate Allele Frequencies from a VCF}
\usage{
vcfR2AF(vcf, popmap, n_cores = 1, allele = "REF")
}
\arguments{
\item{vcf}{a vcf stored as a vcfR object}

\item{popmap}{a two column data.frame, column 1 lists individuals and column 2 population assignment}

\item{n_cores}{number of cores to run with mclapply. If n_cores=1, runs with lapply instead.}

\item{allele}{character vector of either "REF" or "ALT" for which allele frequency to return}
}
\value{
A matrix of population-specific allele frequencies
}
\description{
Takes as input a population map (popmap) and a VCF (vcfR object) and returns an m x n matrix, where m = the number of SNPs and n = the number of populations
}
