% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window2pos_df.R
\name{window2pos_df}
\alias{window2pos_df}
\title{Transform genome windows/co-ordinates to df of chr, start, pos}
\usage{
window2pos_df(genome_windows)
}
\arguments{
\item{genome_windows}{character string or character vector of genome windows/co-ordinates in the format of chr:start-end}
}
\value{
A data.frame of chr, start, end identifiers for genome windows
}
\description{
Takes as input genome windows/co-ordinates in the format of chrX:start-end and splits to 3 column matrix. The function is just a wrapper for tidyr::separate that assumes the chr:start-end format
}
